requirements.txt
pocketflow>=0.0.1
youtube-transcript-api>=0.6.1
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.12.0
watchdog>=6.0.0